Rename actions/misc.js to actions/panelDisplay.js
Renaming the file in preparation for adding more panel display specific
functions in the following commits. There is no other miscellaneous
code in this file, so better to be more specific about its purpose.
