+++ src/actions/panelDisplay.js
